+++ static/boot.js
@@ -1524,6 +1524,7 @@ function applyBotName(){
     window._showTps=!!s.show_tps;
     window._fadeTextEffect=!!s.fade_text_effect;
     window._showCliSessions=!!s.show_cli_sessions;
+    window._showAllWorkspaces=s.show_all_workspaces!==false;  // default true for back-compat (#2549)
     window._showPreviousMessagingSessions=!!s.show_previous_messaging_sessions;
     window._soundEnabled=!!s.sound_enabled;
     window._notificationsEnabled=!!s.notifications_enabled;
@@ -1624,6 +1625,7 @@ function applyBotName(){
     window._showTps=false;
     window._fadeTextEffect=false;
     window._showCliSessions=false;
+    window._showAllWorkspaces=true;
     window._soundEnabled=false;
     window._notificationsEnabled=false;
     window._whatsNewSummaryEnabled=false;